refactor: create layout for auth pages
